# packages/contracts/openapi/enchev-api.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Enchev Auctions HTTP API",
    "version": "0.1.0",
    "description": "Canonical specification of HTTP routes that are implemented in the repository. It is descriptive only and does not create auction authority."
  },
...
  "servers": [
    {
      "url": "https://enchev-auctions.vercel.app",
      "description": "Canonical production web origin"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development origin"
    }
  ],
  "paths": {
...
    "/api/enchev-auth/[action]": {
      "get": {
        "operationId": "getBuyerAuthSession",
        "summary": "Check authenticated ENCHEV buyer session, fail-closed without dedicated provider",
        "parameters": [
          {
            "in": "path",
            "name": "action",
            "required": true,
            "schema": {
              "const": "session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "No-store authentication state, never JWTs or passwords",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestCorrelationId"
              }
            }
          },
          "400": {
            "description": "Unsupported auth action",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestCorrelationId"
              }
            }
          }
        }
      },
      "post": {
        "operationId": "postBuyerAuthAction",
        "summary": "Dedicated-project buyer sign-up, sign-in or sign-out; same-origin only",
        "parameters": [
          {
            "in": "path",
            "name": "action",
            "required": true,
            "schema": {
              "enum": [
                "sign-up",
                "sign-in",
                "sign-out"
              ]
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 12,
                    "maxLength": 128
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Auth outcome with HttpOnly session cookies; credentials never returned",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestCorrelationId"
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestCorrelationId"
              }
            }
          },
          "401": {
            "description": "Bad credentials",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestCorrelationId"
              }
            }
          },
          "403": {
            "description": "Cross-origin POST rejected",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestCorrelationId"
              }
            }
          },
          "429": {
            "description": "Rate-limited by identity provider",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestCorrelationId"
              }
            }
          },
          "503": {
            "description": "Dedicated project not enabled or identity service unavailable",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestCorrelationId"
              }
            }
          }
        }
      }
    }
  },
  "components": {
...
    "headers": {
      "RequestCorrelationId": {
        "description": "Request correlation identifier resolved by the API boundary. Valid inbound X-Request-ID values are preserved; missing or invalid values are replaced.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
        }
      }
    }
  }
}